move executable files (.js files) under cli/ directory
