+++ cli/gulpfile.js
@@ -3,7 +3,7 @@ const gulp = require('gulp');
 const del = require('del');
 const chalk = require('chalk');
 const gulpChange = require('gulp-change');
-const paths = require('./build.json');
+const paths = require('../config/build.json');
 const plugins = require('gulp-load-plugins')();
 const webpack = require('webpack-stream');
 const sass = require('gulp-sass')(require('sass'));
