+++ cli/xt-create.js
@@ -21,8 +21,8 @@ const exec = require('child_process').exec;
 const Spinner = require('cli-spinner').Spinner;
 const spinner = new Spinner(' %s ');
 const Utilities = require('./utilities').Utilities;
-const texts = require('../config/texts').xtCreate;
-const createPrompts = require('../config/createPrompts').prompts;
+const texts = require('./texts').xtCreate;
+const createPrompts = texts.prompts;
 const defaultHomepage = 'http://chrome.google.com/webstore'
 const initFilesPath = '../config/init/';
