+++ cli/xt-docs.js
@@ -34,7 +34,7 @@ const pkg = require('../package.json');
 const exec = require('child_process').exec;
 const Spinner = require('cli-spinner').Spinner;
 const Utilities = require('./utilities').Utilities;
-const texts = require('../config/texts').xtDocs;
+const texts = require('./texts').xtDocs;
 const defaultConfig = require('../config/docs.json');
 const spinner = new Spinner(' %s ');
 const jsdoc = './node_modules/.bin/jsdoc';
